+++ etils/ecolab/inspects/resource_utils.py
@@ -22,7 +22,7 @@ from etils import epath
 
 def _static_path() -> epath.Path:
   """Path to the static resources (`.js`, `.css`)."""
-  return epath.resource_path('etils') / 'ecolab' / 'inspects' / 'static'
+  return epath.resource_path('etils.ecolab') / 'inspects' / 'static'
 
 
 # TODO(epot): Use gstatic to serve those files.